Add vdp command server
It is a waste of time waiting the end of command while we can do
other thing. The best solution is keep a queue of commands,
and check some times if there is some command ready to run.
It is very important don't forgive check the state, because in other
case you will not have your vdp commands.