Fixed bug in SPRITE function
Sprite patterns in v9938 are 8x8, so the size of a sprite pattern is 8
bytes, but when you use the sprite mode 16x16 the patterns are joined in
groups of 4 patterns, and you have to use the pattern number and not the
pattern group number. The code was doing calculations for groups (32 bytes)
instead of doing the calculations for individual patterns.