Clean the low part of the pattern page
This page can contains some other graphics that are not important for us
so the best solution is clean this part of the page. Like we want to clean
we have to use color 0, and this can cause conflict with the grid color,
so we define a color for the grid.