Convert ISOREGION into a callback function
Having a function which runs over a region and calls a callback function
is better than have a specialized version, because we will need it later
(for example in the moment we want to paint the tile instead of marking it).

Like this is something we will use later in the game, it is not logical put it
in the editor.asm source file, so it is added a new file, isometri.asm.