Copy the selected tile to selected region.
The user selects the tile he want to use and later selects the region which
he wants to fill. In that moment it is necesarry call to ISORECT with a new
callback function, the function which will copy the tiles.

After the copy, if you write the grid again you will overwrite the tile, so
it is necessary move the call to GRID out of the editor main loop.