Move to function the question about visible tile
When we are operating in a region, maybe some elements of the region are
outside of the screen, and in this case we don't have to operate in them.
The best place should be in the loop which runs over the region, and don't
worry about it in the callback function, but these callback functions can be
called out of the loop, so it is necessary they check the visibility of the
current tile.