Simplify DRAWREGION
DRAWREGION was too much complex. It could paint lines (horizontal and vertical)
and cut all the edges of a isometric round. This is not necessary in DRAWREGION
because vertical regions (they come soon ;)) can do it better (they are 2
masked lines around usual lines, and all of them are pararells).
After this commit a region with x or y = 0 hasn't any effect.