Let RECTANGLE modify ISODIR
RECTANGLE worked fine only when the value of ISODIR was correct set with the
value of ISOTYPE. Like RECTANGLE was tested in the function, it is better
let that RECTANGLE modify it and it will work always. If you have to use
lines and rectangles, set the value you need for the lines and draw it, and
later draw the rectangles, and you will not have problems.