Use only one sprite as pointer
The code was using 1 sprite plane for pattern selecting and other for tile
selecting, but they were never used at same time, so it is better use only
one sprite plane for both things.
In previous commits we have added the ISOTILE variable which indicate if we
are using a region or tile command, so it is good idea show different
pointer in such cases (This task is done by MARKTILE of EDITOR.ASM).