Remove EDKEV function
EDKEV was an attempt of having a centralized function which could handle
all the common keys of the different function in editor.asm, but it was not
working because it couldn't remove some actions once it was called (for
example remove the selection for a region or a meta pattern). So, in this case
is better dealing with the keyboard in each function (the number of functions
is enough small for this aproach).