Use CTRL for moving the region in euclidian movement
You can move the initial tile for the regions using SHIFT, but it does it
using a isometric movement, that not always is that user needs. This patch adds
CTRL for moving the initial tile using euclidian movement. This makes sense
because user can move the pointer in isometric way with SHIFT, so the same
key is used for the same in both context