Add PUTMOB function
A mob is a Movable Object Block, and it is an implementation of software
sprites. It is used the name 'mob' in order to avoid the confusion of
sprite. In order to can test the code it is also added a non functional game
module. which only copy the content of the TILPAGE to the second page and
the background page. This code will be modified later.