Disable interrupts in the vram transfers
The vdp hook send the sprite attribute table to the vdp in each interrupt,
it means that each interrupt the vram pointer is modified, so a function
can't assume that it is constant. The function which write/read directly
to/from vram are used only in initialization, so it is not a problem disable
interrupts in vram transfer. This solution fixes a bug which only appears
in z80 mode, where the interrupts were fired inside the vram transfer of
sprite initialization and they were not correctly displayed.