Remove PRINTNUM function in EDITOR.ASM
This function was printing the pattern number in the pattern selection mode,
ut thisinformation was no util at all, so this patch removes it. This
function used a BCD function that convert from natural binary to string,
that surely will be used in the future, so a new file called stdlib is
created where we also put the memset code. This new file is created
for all the C alike functions that the code could need.