Use GEOMETRY functions in command selection
Selecting the command means increment or decrement a value in base a
directional key, so basically is the same that in other case. The unique
difference in this case is up/down actions are reversed, so it is necessary
add a new movement funciton in geometry.
This patch also undelete the sprites in the command selection, so it is
easier see what you are doing.