Add Z buffer to the mapper routines
The Z buffer allow us know the value of the Z coordenate of each pattern
rendered in the same pattern. In order to can do it we use a sparse matrix
where each element of the matrix is a array of NR_PATTIL elements, having
each one of them the z value.