Split world movement and screen movement
It is a good idea having world movement not related to screen movement. World
movement is done in a usual matrix, and it is an euclidean movement of 1x1,
but we want a small animation in the screen, so each world movement is
translated to 4 small movement in the screen.