Unmark regions only when it is needed
The loop in MARKREGION was:

    CALL MARKREGION
    CALL GETKEV
    CALL MARKREGION

getting that each key pressed mark the region 4 times, 2 times for the press
and 2 times for the release. This behaviour was causing that marking big
regions what too much slow. This patch fix the problem marking the region
only when a directional key is pressed.