Allow negative values for MAPINC
MAPINC holds the value used as increment of (TILE) in command repetions, and
we were packing it in a single byte. After depacking it, the allowed values
for MAPINC were between 0 and 16. It was very frustating because it is needed
negative values, so this patch adds negative numbers for MAPINC.