Fix deletion of repeat commands
Delete was using CRUNCH in order to know the size of last command, and REPEAT
commands were detected reading the MAPREP variable, but in order to avoid
next call to REPEAT, this variable was reset after the command, so new calls
to CRUNCH return wrong information. This patch fixes the problem using a
internal variable which holds the last MAPCMD executed.