Fix MOB background restoring
The MOB engine was drawing all the mobs and later restoring the background.
Using this algorithm two overlapped mobs couldn't restore the background,
because restoring the background of first delete the second mob. This patch
restores the background as soon is possible (just drawing the mob), so the
problem is fixed.