Rewrite keyboard routine
This new version uses the BIOS variables instead of reading directly
from the PPI. It is also changed the names of the routines because could
cause some confusion (GETCHAR was not a valid name because we are working with
scancodes and not ascii codes). All the RAM variables are located in work area
due to bios ISR changed the pages, so using work area we can be sure the
data is accesible.