Check room location is correct
We will need check room at left, or rigth of some other room, and not
always they will be a valid room. Other parts of the code will check
this issue and mark them as invalid with -1,-1. In this case GETROOM and
ROOM2HGT and ROOM2MAP map check against this case.