Check out of screen in REMAP
REMAP has to check if the region is not aligned with the patterns, because in
this case is necessary copy one pattern more. The problem comes when we
are in the limits of the screen, because in this case this increment means go
far of the screen, and check a tile stack which is out of the real data
structure.