Add TURNON and TURNOFF
These functions swith off, or swith on the screen visualization. They must be
called before and after map, because they claen/set the vram content in the
way the mob engine expects. This is the reason these functions are located in
mob.asm, because they are too much related with the mob implementation.