Hide movables in when change camera
TURNOFF was freeing all the mobs, but this caused a problem, because the
movables which own these mobs didn't know anything about it. The solution
must be remove the freeing from TURNOFF and add a new function in movable.asm
which runs over all the movables and frees the mobs asocciated to them. This
function must be called after TURNOFF after a camera change.