Move program segments
Due to the program we were using for debugging, actual slot configuration
is not correct, because we change the page 2, and in that page is where
debugger set the stack, so debugging process is paintful. This patch change
the page configuration to:

  - code: pagina 0
  - levels: pagina 1
  - data: pagina 2

Like the code is in page 0, we can not set the hook jump there, so it is
necessary move the hook routine to the data page and change the slot before
calling our functions.