Move SHOWHEIGTHS to editor.asm
SHOWGRID AND SHOWLIMITS are in editor.asm, so it is logical SHOWHEIGTHS be in
editor.asm to. It helps to have it private and not public like it was in
edheigths.asm. It is also splitted ED.SCREEN, letting the non editor part
in DBGSCR (debug screen), so we can use it in other part.