Use different keys to launc each editor
Level editor after space pressing launched first the map editor and after
it the heigth editor. It caused it was necessary pass always over the map
editor in order to can run the heigth editor. This patch removes this
ugly behaviour, setting F1 for map editor and F2 for heigth editor.