Rewrite level.asm
This rewrited version used a centralized data structure, where you can
find all the data of a room, since the increment from central room, heigth
matrix (not a pointer to the matrix), and the own room location. Having such
structure allows having a pointer in the movable to it, getting easily
all the information related to the room (and it is necessary calculate it
only once).