Add heigth to movables
When movables move in the game, the heigths of the world are modified, so
it is important update the heigth matrix when a movable is placed in some
place. The perfect function for this yask is SETRINFO, because is the
place where the room info is got. It is necessary check if we are in a
animation, because in this case we also have to update heigth of destination.