Remove MOV16ISO function
This function returned the next position doing a isometric increment. This
solution was a problem because it means store the render coordenates in the
movable in each animation step, but since the camera can change in the middle
of the animation, it can cause incorrect data. We need a function that
based in a direction and a number of animation steps, it returns the increment
from the base position, and thank to it, we can change the base position
without problems.