Remove buffers from work area
Long time ago we had a problem because we put the data in page 0,
and when interupt arrived, then we don't have this page. Solution was
putting variables used in the isr in the work area. Our actual ISR set
the pages configuration we want, so using work area is no longer needed. This
patch moves all the buffers to the data segment, and it helps us to keep them
private.