Remove page byte
All the pages have the first byte with the page number. It was good because
it simplifies the loader, but it has the problem you can not load a 256x128
gl5, and it is the format the tile set will have. The solution is use a
hardcoded loader for each page.