Add delgl5 tool
This tool allows remove the first 4 bytes of a gl5 file. This 4 bytes are the
x,y size of the copy, and they are good. But in the case of tileset, we will
store 256x128 size copy, and they fit in a 16K page, but without the 4 bytes.
The fist solution was using a c program, but duew to the use of CPM calls the
file after removing the 4 bytes was bigger. So the unique solution was using
an assembler program with DOS2 calls (why the damm I can not use a good UNIX
implementation in MSX!!!!!).