Fix test conditions in map editor
The code was testing if map number was 0, and in this case it didn't let
modify the map number, but it was incorrect, because it caused any room could
be modified in the beginning (all the rooms have map 0). The correct test is
don't let modify the map content when map number is equal to 0.