Fix hexadecimal format in printf
Hexadecimal was not working because HEX waits a buffer parameter and
nothing was passed. There were also some problems with the conversion to
ascii, because PRINTNUM waits a 0 in the end of the string, and nothing was
passed. It is also modified the main loop in printf, because there were too
much push/pop.