Remove magic numbers in the code
We were using a lot of magic numbers in the code, and in some places
we were using a symbolic name for more of one thing. Although we know
they must have the same value it is better use different symbolic names
because it increases the readibility of the code.