Add HMATRIX in RINFO
When a room is showed all the information is cached in the RINFO struct. If
the room is not showed we don't need check heights, so we can keep them
in the rom, but in the showed rooms we also need update the height due to
the movables, so we need a ram copy of them.