Rename PAT2XY to PNUM2XY and add a real PAT2XY
We are calling a pattern the pattern coordinates in the screen, and the
number which represents the pattern is the pattern number. Call PAT2XY to the
function which transform from the pattern number to the screen coordinates
of the bitmap causes confusion. It is the reason why we transform it to
PNUM2XY and we add the real PAT2XY.