Move pattern adjust to WRDLS2SCR
We have to do a coordenates transform in PUTMOB, because the movable module
uses the floor address, but in mob module we need the upper corner of the
copy. WRLD2SCR returned the upper corner of the pattern, but we need the lower
corner. PUTMOB was doing an adjust due to this problem.