Generate height information
The height information is needed in order can detect forbidden parts of
the map, and parts where are stairs, so this information is essential. We have
to put the button in the edmap because we need know the height of the map,
and this information is not stored in the own map, it is part of the room
information.