Add empty floor
We need a floor if we want a position can be accesable, because in other case
it will be a hole to lower level. Some floors will not be drawable never,
so it is a waste of time paint something that will no be visible. The
best solution is define a new floor without graphic that canbe used in this
way.