Fix bug in GETHEIGHT
GETHEIGHT returns a pointer to the height position where a movable is located.
It is needed by STEP that these position can be modified, and we can not
modify the values which are stored in ROM, so the solution is using a dummy
position for this modfications. This was done in the previous version of
GETHEIGHT, but it was removed in this last.