Update level in NEXTPOINT
NEXTPOINT is used in step, and in this moment is not modifing the level part
of the coordenates, and it causes that is not possible pass from a level to
other level. This patch add this possibility, but it is not possible render
both levels at the same time, so it is visible the break between both levels.