Add /T to copys in MAKEFILE
It seems that the make that we are using only checks
if the time is derivates is estrictly bigger, and it
means that it always rebuild targets with prerequisites
with the same time.

DOS always copy the time of the source file, but it
can be avoided with the /T option.